+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0
+
+- Rename a few methods
+  - `make-client` -> `build-client`
+  - `make-request` -> `build-request`
+  - `resp->ring` -> `response->map`
+
 ## 0.1.0
 
 - Initial release